APP/JudgeEverything/Copilot/python/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
opencv-python
torch
torchvision
python-multipart
aiofiles
requests
base64
